feature/sdk_logger - update application code
