+++ Apester/AppDelegate.swift
@@ -58,60 +58,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
     
 }
-public class DemoLogWriter: APELogWriter {
-    
-    private let stateLock = DemoThreadMutex()
-    private var _entries: [APELogger.Entry] = []
-    
-    public init() { }
-    
-    @discardableResult
-    private func synchronise<T>(block: () -> T) -> T {
-        return stateLock.withCriticalScope(block: block)
-    }
-    
-    public var entries: [APELogger.Entry] {
-        return synchronise { _entries }
-    }
-    
-    public func write(entry: APELogger.Entry) {
-        synchronise {
-            _entries.append(entry)
-        }
-    }
-}
-/// A wrapper class for a pthread_mutex
-internal final class DemoThreadMutex {
-    private var mutex = pthread_mutex_t()
-    
-    public init() {
-        let result = pthread_mutex_init(&mutex, nil)
-        precondition(result == 0, "Failed to create pthread mutex")
-    }
-    
-    deinit {
-        let result = pthread_mutex_destroy(&mutex)
-        assert(result == 0, "Failed to destroy mutex")
-    }
-    
-    fileprivate func lock() {
-        let result = pthread_mutex_lock(&mutex)
-        assert(result == 0, "Failed to lock mutex")
-    }
-    
-    fileprivate func unlock() {
-        let result = pthread_mutex_unlock(&mutex)
-        assert(result == 0, "Failed to unlock mutex")
-    }
-    
-    /// Convenience API to execute block after acquiring the lock
-    ///
-    /// - Parameter block: the block to run
-    /// - Returns: returns the return value of the block
-    public func withCriticalScope<T>(block: () -> T) -> T {
-        lock()
-        defer { unlock() }
-        let value = block()
-        return value
-    }
-}
